Move RwContext to rwContext.js.
