+++ sdk/src/runtime/lib/rwContext.ts
@@ -1,0 +1,25 @@
+import { type RequestInfo } from "../requestInfo/types.js";
+import { type Kysely } from "kysely";
+
+export type RwContext = {
+  nonce: string;
+  Document: React.FC<DocumentProps<any>>;
+  rscPayload: boolean;
+  ssr: boolean;
+  layouts?: React.FC<LayoutProps<any>>[];
+  databases: Map<string, Kysely<any>>;
+  scriptsToBeLoaded: Set<string>;
+  entryScripts: Set<string>;
+  inlineScripts: Set<string>;
+  pageRouteResolved: PromiseWithResolvers<void> | undefined;
+  actionResult?: unknown;
+};
+
+export type DocumentProps<T extends RequestInfo = RequestInfo> = T & {
+  children: React.ReactNode;
+};
+
+export type LayoutProps<T extends RequestInfo = RequestInfo> = {
+  children?: React.ReactNode;
+  requestInfo?: T;
+};
